msp: restructure to export stacks implementations
