+++ dev/managedservicesplatform/BUILD.bazel
@@ -10,15 +10,15 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//dev/managedservicesplatform/internal/stack",
-        "//dev/managedservicesplatform/internal/stack/cloudrun",
-        "//dev/managedservicesplatform/internal/stack/iam",
-        "//dev/managedservicesplatform/internal/stack/monitoring",
         "//dev/managedservicesplatform/internal/stack/options/terraformversion",
         "//dev/managedservicesplatform/internal/stack/options/tfcbackend",
-        "//dev/managedservicesplatform/internal/stack/project",
-        "//dev/managedservicesplatform/internal/stack/tfcworkspaces",
         "//dev/managedservicesplatform/internal/terraform",
         "//dev/managedservicesplatform/spec",
+        "//dev/managedservicesplatform/stacks/cloudrun",
+        "//dev/managedservicesplatform/stacks/iam",
+        "//dev/managedservicesplatform/stacks/monitoring",
+        "//dev/managedservicesplatform/stacks/project",
+        "//dev/managedservicesplatform/stacks/tfcworkspaces",
         "//dev/managedservicesplatform/terraformcloud",
         "//lib/errors",
         "//lib/pointers",
