+++ dev/managedservicesplatform/stacks/README.md
