+++ dev/managedservicesplatform/stacks/cloudrun/BUILD.bazel
@@ -3,7 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "cloudrun",
     srcs = ["cloudrun.go"],
-    importpath = "github.com/sourcegraph/sourcegraph/dev/managedservicesplatform/internal/stack/cloudrun",
+    importpath = "github.com/sourcegraph/sourcegraph/dev/managedservicesplatform/stacks/cloudrun",
     visibility = ["//dev/managedservicesplatform:__subpackages__"],
     deps = [
         "//dev/managedservicesplatform/googlesecretsmanager",
@@ -17,15 +17,15 @@ go_library(
         "//dev/managedservicesplatform/internal/resource/tfvar",
         "//dev/managedservicesplatform/internal/resourceid",
         "//dev/managedservicesplatform/internal/stack",
-        "//dev/managedservicesplatform/internal/stack/cloudrun/internal/builder",
-        "//dev/managedservicesplatform/internal/stack/cloudrun/internal/builder/job",
-        "//dev/managedservicesplatform/internal/stack/cloudrun/internal/builder/service",
-        "//dev/managedservicesplatform/internal/stack/iam",
         "//dev/managedservicesplatform/internal/stack/options/cloudflareprovider",
         "//dev/managedservicesplatform/internal/stack/options/dynamicvariables",
         "//dev/managedservicesplatform/internal/stack/options/googleprovider",
         "//dev/managedservicesplatform/internal/stack/options/randomprovider",
         "//dev/managedservicesplatform/spec",
+        "//dev/managedservicesplatform/stacks/cloudrun/internal/builder",
+        "//dev/managedservicesplatform/stacks/cloudrun/internal/builder/job",
+        "//dev/managedservicesplatform/stacks/cloudrun/internal/builder/service",
+        "//dev/managedservicesplatform/stacks/iam",
         "//lib/errors",
         "//lib/pointers",
         "@com_github_hashicorp_terraform_cdk_go_cdktf//:cdktf",
