+++ dev/managedservicesplatform/stacks/cloudrun/internal/builder/service/BUILD.bazel
@@ -3,7 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "service",
     srcs = ["service.go"],
-    importpath = "github.com/sourcegraph/sourcegraph/dev/managedservicesplatform/internal/stack/cloudrun/internal/builder/service",
+    importpath = "github.com/sourcegraph/sourcegraph/dev/managedservicesplatform/stacks/cloudrun/internal/builder/service",
     visibility = ["//dev/managedservicesplatform:__subpackages__"],
     deps = [
         "//dev/managedservicesplatform/internal/resource/cloudflare",
@@ -11,8 +11,8 @@ go_library(
         "//dev/managedservicesplatform/internal/resource/loadbalancer",
         "//dev/managedservicesplatform/internal/resource/managedcert",
         "//dev/managedservicesplatform/internal/resourceid",
-        "//dev/managedservicesplatform/internal/stack/cloudrun/internal/builder",
         "//dev/managedservicesplatform/spec",
+        "//dev/managedservicesplatform/stacks/cloudrun/internal/builder",
         "//lib/errors",
         "//lib/pointers",
         "@com_github_hashicorp_terraform_cdk_go_cdktf//:cdktf",
