+++ dev/managedservicesplatform/stacks/iam/BUILD.bazel
@@ -4,7 +4,7 @@ load("//dev:go_defs.bzl", "go_test")
 go_library(
     name = "iam",
     srcs = ["iam.go"],
-    importpath = "github.com/sourcegraph/sourcegraph/dev/managedservicesplatform/internal/stack/iam",
+    importpath = "github.com/sourcegraph/sourcegraph/dev/managedservicesplatform/stacks/iam",
     visibility = ["//dev/managedservicesplatform:__subpackages__"],
     deps = [
         "//dev/managedservicesplatform/internal/resource/random",
