+++ dev/managedservicesplatform/stacks/monitoring/BUILD.bazel
@@ -3,7 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "monitoring",
     srcs = ["monitoring.go"],
-    importpath = "github.com/sourcegraph/sourcegraph/dev/managedservicesplatform/internal/stack/monitoring",
+    importpath = "github.com/sourcegraph/sourcegraph/dev/managedservicesplatform/stacks/monitoring",
     visibility = ["//dev/managedservicesplatform:__subpackages__"],
     deps = [
         "//dev/managedservicesplatform/googlesecretsmanager",
