+++ dev/managedservicesplatform/stacks/project/BUILD.bazel
@@ -3,7 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "project",
     srcs = ["project.go"],
-    importpath = "github.com/sourcegraph/sourcegraph/dev/managedservicesplatform/internal/stack/project",
+    importpath = "github.com/sourcegraph/sourcegraph/dev/managedservicesplatform/stacks/project",
     visibility = ["//dev/managedservicesplatform:__subpackages__"],
     deps = [
         "//dev/managedservicesplatform/internal/resourceid",
